+++ Foundation/Foreign.hs
@@ -9,11 +9,11 @@ module Foundation.Foreign
     ( module Basement.FinalPtr
     , V.foreignMem
     , V.mutableForeignMem
-    , module Foreign.C.Types
+    , module Basement.Compat.C.Types
     ) where
 
 import           Basement.FinalPtr
 import qualified Basement.UArray as V
 import qualified Basement.UArray.Mutable as V
 
-import           Foreign.C.Types
+import           Basement.Compat.C.Types