+++ Foundation/System/Bindings/Hs.hs
@@ -5,6 +5,6 @@ module Foundation.System.Bindings.Hs
     where
 
 import GHC.IO
-import Foreign.C.Types
+import Basement.Compat.C.Types
 
 foreign import ccall unsafe "HsBase.h __hscore_get_errno" sysHsCoreGetErrno :: IO CInt