+++ basement/Basement/Block/Mutable.hs
@@ -38,8 +38,11 @@ module Basement.Block.Mutable
     ( Block(..)
     , MutableBlock(..)
     , mutableLengthSize
+    , mutableLength
     , mutableLengthBytes
     , mutableWithPtr
+    , withMutablePtr
+    , withMutablePtrHint
     , new
     , newPinned
     , mutableEmpty
@@ -56,11 +59,15 @@ module Basement.Block.Mutable
     , unsafeCopyBytes
     , unsafeCopyBytesRO
     , unsafeCopyBytesPtr
+    -- * Foreign
+    , copyFromPtr
+    , copyToPtr
     ) where
 
 import           GHC.Prim
 import           GHC.Types
 import           Basement.Compat.Base
+import           Basement.Compat.Primitive (compatCopyByteArrayToAddr#)
 import           Data.Proxy
 import           Basement.Exception
 import           Basement.Types.OffsetSize
@@ -69,20 +76,6 @@ import           Basement.Numerical.Additive
 import           Basement.PrimType
 import           Basement.Block.Base
 
--- | Return the length of a Mutable Block
---
--- note: we don't allow resizing yet, so this can remain a pure function
-mutableLengthSize :: forall ty st . PrimType ty => MutableBlock ty st -> CountOf ty
-mutableLengthSize (MutableBlock mba) =
-    let !(CountOf (I# szBits)) = primSizeInBytes (Proxy :: Proxy ty)
-        !elems              = quotInt# (sizeofMutableByteArray# mba) szBits
-     in CountOf (I# elems)
-{-# INLINE[1] mutableLengthSize #-}
-
-mutableLengthBytes :: MutableBlock ty st -> CountOf Word8
-mutableLengthBytes (MutableBlock mba) = CountOf (I# (sizeofMutableByteArray# mba))
-{-# INLINE[1] mutableLengthBytes #-}
-
 -- | Set all mutable block element to a value
 iterSet :: (PrimType ty, PrimMonad prim)
         => (Offset ty -> ty)
@@ -90,11 +83,15 @@ iterSet :: (PrimType ty, PrimMonad prim)
         -> prim ()
 iterSet f ma = loop 0
   where
-    !sz = mutableLengthSize ma
+    !sz = mutableLength ma
     loop i
         | i .==# sz = pure ()
         | otherwise = unsafeWrite ma i (f i) >> loop (i+1)
     {-# INLINE loop #-}
+
+mutableLengthSize :: PrimType ty => MutableBlock ty st -> CountOf ty
+mutableLengthSize = mutableLength
+{-# DEPRECATED mutableLengthSize "use mutableLength" #-}
 
 -- | read a cell in a mutable array.
 --
@@ -103,7 +100,7 @@ read :: (PrimMonad prim, PrimType ty) => MutableBlock ty (PrimState prim) -> Off
 read array n
     | isOutOfBound n len = primOutOfBound OOB_Read n len
     | otherwise          = unsafeRead array n
-  where len = mutableLengthSize array
+  where len = mutableLength array
 {-# INLINE read #-}
 
 -- | Write to a cell in a mutable array.
@@ -116,3 +113,42 @@ write array n val
   where
     len = mutableLengthSize array
 {-# INLINE write #-}
+
+-- | Copy from a pointer, @count@ elements, into the Mutable Block at a starting offset @ofs@
+--
+-- if the source pointer is invalid (size or bad allocation), bad things will happen
+--
+copyFromPtr :: forall prim ty . (PrimMonad prim, PrimType ty)
+            => Ptr ty                           -- ^ Source Ptr of 'ty' to start of memory
+            -> MutableBlock ty (PrimState prim) -- ^ Destination mutable block
+            -> Offset ty                        -- ^ Start offset in the destination mutable block
+            -> CountOf ty                       -- ^ Number of 'ty' elements
+            -> prim ()
+copyFromPtr src@(Ptr src#) mb@(MutableBlock mba) ofs count
+    | end > sizeAsOffset arrSz = primOutOfBound OOB_MemCopy end arrSz
+    | otherwise                = primitive $ \st -> (# copyAddrToByteArray# src# mba od# bytes# st, () #)
+  where
+    end = od `offsetPlusE` arrSz
+
+    sz = primSizeInBytes (Proxy :: Proxy ty)
+    !arrSz@(CountOf (I# bytes#)) = sizeOfE sz count
+    !od@(Offset (I# od#)) = offsetOfE sz ofs
+
+-- | Copy all the block content to the memory starting at the destination address
+--
+-- If the destination pointer is invalid (size or bad allocation), bad things will happen
+copyToPtr :: forall ty prim . (PrimType ty, PrimMonad prim)
+          => MutableBlock ty (PrimState prim) -- ^ The source mutable block to copy
+          -> Offset ty                        -- ^ The source offset in the mutable block
+          -> Ptr ty                           -- ^ The destination address where the copy is going to start
+          -> CountOf ty                       -- ^ The number of bytes
+          -> prim ()
+copyToPtr mb@(MutableBlock mba) ofs dst@(Ptr dst#) count
+    | srcEnd > sizeAsOffset arrSz = primOutOfBound OOB_MemCopy srcEnd arrSz
+    | otherwise                = do
+        (Block ba) <- unsafeFreeze mb
+        primitive $ \s1 -> (# compatCopyByteArrayToAddr# ba os# dst# szBytes# s1, () #)
+  where
+    srcEnd = os `offsetPlusE` arrSz
+    !os@(Offset (I# os#)) = offsetInBytes ofs
+    !arrSz@(CountOf (I# szBytes#)) = mutableLengthBytes mb