+++ basement/Basement/UArray/Base.hs
@@ -61,6 +61,7 @@ import           Basement.Compat.Primitive
 import           Basement.Monad
 import           Basement.PrimType
 import           Basement.Compat.Base
+import           Basement.Compat.C.Types
 import           Basement.Compat.Semigroup
 import qualified Basement.Runtime as Runtime
 import           Data.Proxy
@@ -74,7 +75,6 @@ import qualified Basement.Block as BLK
 import qualified Basement.Block.Mutable as MBLK
 import           Basement.Numerical.Additive
 import           Basement.Bindings.Memory
-import           Foreign.C.Types
 import           System.IO.Unsafe (unsafeDupablePerformIO)
 
 -- | A Mutable array of types built on top of GHC primitive.
@@ -171,6 +171,10 @@ isMutablePinned (MUArray _ _ (MUArrayMBA mb))  = BLK.isMutablePinned mb
 newPinned :: forall prim ty . (PrimMonad prim, PrimType ty) => CountOf ty -> prim (MUArray ty (PrimState prim))
 newPinned n = MUArray 0 n . MUArrayMBA <$> MBLK.newPinned n
 
+-- | Create a new unpinned mutable array of size @n elements.
+--
+-- If the size exceeds a GHC-defined threshold, then the memory will be
+-- pinned. To be certain about pinning status with small size, use 'newPinned'
 newUnpinned :: forall prim ty . (PrimMonad prim, PrimType ty) => CountOf ty -> prim (MUArray ty (PrimState prim))
 newUnpinned n = MUArray 0 n . MUArrayMBA <$> MBLK.new n